+++ test/mailers/devise_mailer_test.rb
@@ -1,89 +1,78 @@
 # frozen_string_literal: true
 
-require 'test_helper'
+require "test_helper"
 
 # Test to make sure devise emails generate correctly
 class DeviseMailerTest < ActionMailer::TestCase
-  test 'confirmation instructions email for applicant' do
+  test "confirmation instructions email for applicant" do
     applicant = applicants(:one)
-    email = Devise::Mailer.confirmation_instructions(applicant, 'faketoken').deliver_now
-    assert !ActionMailer::Base.deliveries.empty?
-    assert_equal [applicant.email], email.to
-    assert_equal 'Confirmation instructions', email.subject
-    assert_match(%r{#{ENV['website_url']}/trainee/confirmation\?confirmation_token=faketoken}, email.encoded)
+    mail = Devise::Mailer.confirmation_instructions(applicant, "faketoken")
+    assert_equal [applicant.email], mail.to
+    assert_equal "Confirm your Train Tracks email address", mail.subject
+    assert_match(%r{#{ENV["website_url"]}/trainee/confirmation\?confirmation_token=faketoken}, mail.body.encoded)
   end
 
-  test 'confirmation instructions email for preceptor' do
+  test "confirmation instructions email for preceptor" do
     preceptor = preceptors(:one)
-    email = Devise::Mailer.confirmation_instructions(preceptor, 'faketoken').deliver_now
-    assert !ActionMailer::Base.deliveries.empty?
-    assert_equal [preceptor.email], email.to
-    assert_equal 'Confirmation instructions', email.subject
-    assert_match(%r{#{ENV['website_url']}/preceptor/confirmation\?confirmation_token=faketoken}, email.encoded)
+    mail = Devise::Mailer.confirmation_instructions(preceptor, "faketoken")
+    assert_equal [preceptor.email], mail.to
+    assert_equal "Confirm your Train Tracks email address", mail.subject
+    assert_match(%r{#{ENV["website_url"]}/preceptor/confirmation\?confirmation_token=faketoken}, mail.body.encoded)
   end
 
-  # test 'confirmation instructions email for user' do
+  # test "confirmation instructions email for user" do
   #   valid = users(:valid)
-  #   email = Devise::Mailer.confirmation_instructions(valid, 'faketoken').deliver_now
-  #   assert !ActionMailer::Base.deliveries.empty?
-  #   assert_equal [valid.email], email.to
-  #   assert_equal 'Confirmation instructions', email.subject
-  #   assert_match(%r{#{ENV['website_url']}/confirmation\?confirmation_token=faketoken}, email.encoded)
+  #   mail = Devise::Mailer.confirmation_instructions(valid, "faketoken")
+  #   assert_equal [valid.email], mail.to
+  #   assert_equal "Confirmation instructions", mail.subject
+  #   assert_match(%r{#{ENV["website_url"]}/confirmation\?confirmation_token=faketoken}, mail.body.encoded)
   # end
 
-  test 'reset password email for applicant' do
+  test "reset password email for applicant" do
     applicant = applicants(:one)
-    email = Devise::Mailer.reset_password_instructions(applicant, 'faketoken').deliver_now
-    assert !ActionMailer::Base.deliveries.empty?
-    assert_equal [applicant.email], email.to
-    assert_equal 'Reset password instructions', email.subject
-    assert_match(%r{#{ENV['website_url']}/trainee/password/edit\?reset_password_token=faketoken}, email.encoded)
+    mail = Devise::Mailer.reset_password_instructions(applicant, "faketoken")
+    assert_equal [applicant.email], mail.to
+    assert_equal "Reset password for your Train Tracks account", mail.subject
+    assert_match(%r{#{ENV["website_url"]}/trainee/password/edit\?reset_password_token=faketoken}, mail.body.encoded)
   end
 
-  test 'reset password email for preceptor' do
+  test "reset password email for preceptor" do
     preceptor = preceptors(:one)
-    email = Devise::Mailer.reset_password_instructions(preceptor, 'faketoken').deliver_now
-    assert !ActionMailer::Base.deliveries.empty?
-    assert_equal [preceptor.email], email.to
-    assert_equal 'Reset password instructions', email.subject
-    assert_match(%r{#{ENV['website_url']}/preceptor/password/edit\?reset_password_token=faketoken}, email.encoded)
+    mail = Devise::Mailer.reset_password_instructions(preceptor, "faketoken")
+    assert_equal [preceptor.email], mail.to
+    assert_equal "Reset password for your Train Tracks account", mail.subject
+    assert_match(%r{#{ENV["website_url"]}/preceptor/password/edit\?reset_password_token=faketoken}, mail.body.encoded)
   end
 
-  test 'reset password email for user' do
+  test "reset password email for user" do
     valid = users(:valid)
-    email = Devise::Mailer.reset_password_instructions(valid, 'faketoken').deliver_now
-    assert !ActionMailer::Base.deliveries.empty?
-    assert_equal [valid.email], email.to
-    assert_equal 'Reset password instructions', email.subject
-    assert_match(%r{#{ENV['website_url']}/password/edit\?reset_password_token=faketoken}, email.encoded)
+    mail = Devise::Mailer.reset_password_instructions(valid, "faketoken")
+    assert_equal [valid.email], mail.to
+    assert_equal "Reset password for your Train Tracks account", mail.subject
+    assert_match(%r{#{ENV["website_url"]}/password/edit\?reset_password_token=faketoken}, mail.body.encoded)
   end
 
-  test 'unlock instructions email for applicant' do
+  test "unlock instructions email for applicant" do
     applicant = applicants(:one)
-    email = Devise::Mailer.unlock_instructions(applicant, 'faketoken').deliver_now
-    assert !ActionMailer::Base.deliveries.empty?
-    assert_equal [applicant.email], email.to
-    assert_equal 'Unlock instructions', email.subject
-    assert_match(%r{#{ENV['website_url']}/trainee/unlock\?unlock_token=faketoken}, email.encoded)
+    mail = Devise::Mailer.unlock_instructions(applicant, "faketoken")
+    assert_equal [applicant.email], mail.to
+    assert_equal "Unlock your Train Tracks account", mail.subject
+    assert_match(%r{#{ENV["website_url"]}/trainee/unlock\?unlock_token=faketoken}, mail.body.encoded)
   end
 
-  test 'unlock instructions email for preceptor' do
+  test "unlock instructions email for preceptor" do
     preceptor = preceptors(:one)
-    email = Devise::Mailer.unlock_instructions(preceptor, 'faketoken').deliver_now
-    assert !ActionMailer::Base.deliveries.empty?
-    assert_equal [preceptor.email], email.to
-    assert_equal 'Unlock instructions', email.subject
-    assert_match(%r{#{ENV['website_url']}/preceptor/unlock\?unlock_token=faketoken}, email.encoded)
+    mail = Devise::Mailer.unlock_instructions(preceptor, "faketoken")
+    assert_equal [preceptor.email], mail.to
+    assert_equal "Unlock your Train Tracks account", mail.subject
+    assert_match(%r{#{ENV["website_url"]}/preceptor/unlock\?unlock_token=faketoken}, mail.body.encoded)
   end
 
-  # test 'unlock instructions email for user' do
+  # test "unlock instructions email for user" do
   #   valid = users(:valid)
-
-  #   email = Devise::Mailer.unlock_instructions(valid, 'faketoken').deliver_now
-  #   assert !ActionMailer::Base.deliveries.empty?
-
-  #   assert_equal [valid.email], email.to
-  #   assert_equal 'Unlock Instructions', email.subject
-  #   assert_match(%r{#{ENV['website_url']}/unlock\?unlock_token=faketoken}, email.encoded)
+  #   mail = Devise::Mailer.unlock_instructions(valid, "faketoken")
+  #   assert_equal [valid.email], mail.to
+  #   assert_equal "Unlock Instructions", mail.subject
+  #   assert_match(%r{#{ENV["website_url"]}/unlock\?unlock_token=faketoken}, mail.body.encoded)
   # end
 end